Alterações necessárias do banco de dados

+++ database/migrations/2024_03_29_010246_create_saidas_table.php
@@ -16,7 +16,7 @@ class CreateSaidasTable extends Migration
         Schema::create('saidas', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->double('valor');
+            $table->decimal('valor',10,2);
             $table->text('observacao')->nullable();
             $table->bigInteger('user_id')->unsigned();
             $table->bigInteger('id_descricao')->unsigned();
